Rendered feed component to audience view
