+++ src/Audience/Audience.jsx
@@ -1,7 +1,32 @@
 import React, { useState } from 'react'; //eslint-disable-line
+import Feed from '../Shared/Feed.jsx'; //eslint-disable-line
 
-const Audience = ({ profile }) => {
+const samplePerformances = [
+  {
+    name: 'Mickey Mack',
+    location: 'Bushwick',
+    category: 'Freeform Jazz',
+    date: 'Tuesday, Aug. 31',
+    time: '08:30PM',
+  },
+  {
+    name: 'Raymond T. Raymonds',
+    location: 'House of Blues',
+    category: 'Standup Comedy',
+    date: 'Wednesday, Sep. 1',
+    time: '08:30PM',
+  },
+  {
+    name: 'Miss Sky Hawaii',
+    location: 'D.A.R. Constitution Hall',
+    category: 'Rock',
+    date: 'Thursday, Sep. 2nd',
+    time: '08:30PM',
+  },
+];
 
-};
+const Audience = ({ profile }) => (
+  <Feed performances={samplePerformances} />
+);
 
 export default Audience;
